+++ source/Energinet.DataHub.MeteringPoints.Domain/Energinet.DataHub.MeteringPoints.Domain.csproj
@@ -34,8 +34,4 @@ limitations under the License.
     <PackageReference Include="NodaTime" Version="3.0.5" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="MeteringPoints\Consumption" />
-  </ItemGroup>
-
 </Project>
