+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Components/NetSettlementGroup.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components
 {
     public class NetSettlementGroup : EnumerationType
     {
