+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/MarketMeteringPoints/Rules/ConnectionTypeRequirementRule.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Globalization;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
 namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints.Rules
