+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/MarketMeteringPoints/Rules/MeteringMethodDoesNotMatchNetSettlementGroupRuleError.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
