+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/MarketMeteringPoints/Rules/MeteringMethodRule.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
