+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/ChangeMasterData/ConsumptionMeteringPoints/ChangeTests.cs
@@ -18,6 +18,7 @@ using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.MeteringPoints.Application.ChangeMasterData.Consumption;
 using Energinet.DataHub.MeteringPoints.Application.Common;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints;
