+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringPoints/Consumption/ChangeMasterDataTests.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Linq;
 using Energinet.DataHub.MeteringPoints.Domain.Addresses;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Events;
